+++ packages/brick_offline_first_with_graphql/lib/src/graphql_offline_queue_link.dart
@@ -9,22 +9,11 @@ import 'package:logging/logging.dart';
 
 /// Stores all mutation requests in a SQLite database
 class GraphqlOfflineQueueLink extends Link {
-  /// A DocumentNode GraphQL execution interface
-  /// https://pub.dev/documentation/gql_link/latest/link/Link-class.html
-  ///
-  /// Instead of the proscribed [Link] series with the links calling `#forward`
-  /// [_inner] is composed. Storing the request must occur before an HTTPLink
-  /// and validating the request occurred with a SocketException must occur
-  /// after the request. Therefore, `forward` can't be called because it's
-  /// needed on both ends of the request. HTTPLink also doesn't invoke `#forward`
-  /// and can only be used last in a `Link.from([])` invocation.
-  final Link _inner;
-
   final Logger _logger;
 
   final GraphqlRequestSqliteCacheManager requestManager;
 
-  GraphqlOfflineQueueLink(this._inner, this.requestManager)
+  GraphqlOfflineQueueLink(this.requestManager)
       : _logger = Logger('GraphqlOfflineQueueLink#${requestManager.databaseName}');
 
   @override
@@ -39,33 +28,26 @@ class GraphqlOfflineQueueLink extends Link {
       await cacheItem.insertOrUpdate(db, logger: _logger);
     }
 
-    Response response;
-
-    try {
-      // Attempt to make Graphql Request, handle it as a traditional response to do check
-      response = await _inner.request(request).first;
-
+    yield* forward!(request).handleError(
+      (e) async {
+        _logger.warning('#send: $e');
+        final db = await requestManager.getDb();
+        await cacheItem.unlock(db);
+        return;
+      },
+      test: (e) => e is ServerException && e.originalException is SocketException,
+    ).asyncMap((response) async {
       if (response.errors?.isEmpty ?? true) {
         final db = await requestManager.getDb();
         // request was successfully sent and can be removed
         _logger.finest('removing from queue: ${cacheItem.toSqlite()}');
         await cacheItem.delete(db);
       }
-    } on ServerException catch (e) {
-      if (e.originalException is SocketException) {
-        _logger.warning('#send: $e');
-
-        /// When the request is null a generic Graphql error needs to be generated
-        response = Response(errors: [GraphQLError(message: 'Unknown error: $e')], data: null);
-      } else {
-        rethrow;
-      }
-    } finally {
       final db = await requestManager.getDb();
       await cacheItem.unlock(db);
-    }
 
-    yield response;
+      return response;
+    });
   }
 
   /// Parse a request and determines what [OperationType] it is