+++ packages/brick_offline_first_with_graphql/CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Unreleased
 
 * Add `subscribe` for streaming updates of all models
+* Do not use `GraphqlOfflineQueueLink` automatically in Offline First Repository
 
 ## 0.0.1